requirements.txt
fastapi
uvicorn
zeep
lxml
requests
mcp[fastmcp]
python-dotenv
fastapi-mcp
